lint: ruff BLOQUANT — 959 -> 0, jeu de regles fige
`ci.yml` lancait `ruff check tools/ tests/ || true` depuis toujours, avec le
commentaire « a retirer des que la config est en place ». Il n'y avait AUCUNE
section `[tool.ruff]` : ruff tournait avec ses defauts et personne n'avait
jamais lu sa sortie. Le `|| true` est retire, le perimetre etendu a `scripts/`.

CONFIG EXPLICITE plutot qu'heritee — la raison n'est pas cosmetique : le jeu
par defaut de ruff CHANGE d'une version a l'autre, et la CI l'installe SANS
epingle. Sans `select` fige, une mise a jour de l'outil rendrait le lint rouge
du jour au lendemain sur des regles que personne n'a choisies. 17 familles
selectionnees, 22 regles ignorees — chacune avec sa raison en commentaire.

UNE REGRESSION, ET C'EST LA PLUS INTERESSANTE DU LOT.

`isort` a permute deux imports de `tools/match_youtube.py` dont **l'ordre est
porteur** : `from common import …` doit venir en premier parce que son import
declenche le bootstrap qui met la racine du depot sur `sys.path` — sans lui,
`from tools.config.registry import …` n'est pas resolvable en standalone. Le
script ne demarrait plus (`ModuleNotFoundError: No module named 'tools'`).

Deux choses a retenir :
- le commentaire qui documentait la contrainte disait « ci-dessus » et a ete
  DEPLACE AVEC l'import : il decrivait alors l'inverse de ce que faisait le
  code. Un commentaire juste peut devenir faux sans qu'on y touche.
- **aucun lint, aucune relecture de diff n'aurait signale ca** : le code restait
  valide, c'est l'ORDRE qui ne l'etait plus. Seul
  `tests/test_script_invocation.py`, qui lance vraiment chaque script, l'a
  attrape. C'est l'argument pour ce genre de test.
Corrige par `# isort: off` sur le bloc, avec la contrainte ecrite noir sur
blanc. Idempotence verifiee : deux `--fix` d'affilee ne bougent plus rien.

Second piege du meme genre, corrige : `F401` a supprime les re-exports de
`tools/review_render.py` (des noms non utilises DANS le module, mais reimportes
pour que `review_server` les expose aux tests) -> 3 fichiers de test non
collectes, dont ~500 tests. Un `# noqa` sur le BLOC ne tient pas : isort
resepare le bloc et le noqa devient orphelin. Il en faut un PAR LIGNE.

DEUX POINTS OUVERTS, remontes plutot que corriges sous couvert de lint :
- 64 `TRY004` : passer le domaine de `ValueError` a `TypeError` casserait 46
  `except ValueError` dans `tools/`, dont le parser de migration qui les
  convertit en erreurs lisibles. C'est un refactor de la hierarchie
  d'exceptions, a decider pour lui-meme.
- `tools/link_check.py` ouvre des URLs venant des DONNEES du site sans filtrer
  le schema : un `file://` serait lu. Annote POINT OUVERT dans le code, traite
  separement.

Portes : ruff 0 sur tools/ tests/ scripts/, pytest complet vert, couverture
96,82 % statements / 95,97 % branches (garde-fou par metrique en sortie 0),
1826 tests vitest.

+++ tools/make_dispatch.py
@@ -61,7 +61,7 @@ def make_dispatch(source_id: str) -> tuple[int, int, int]:
     pending.sort(key=lambda e: e["date"] or "", reverse=True)
 
     total = len(pending)
-    laptop_n = int(round(total * LAPTOP_SHARE))
+    laptop_n = round(total * LAPTOP_SHARE)
     main_n = total - laptop_n
 
     # Pour équilibrer le ressenti, la main fait des épisodes RÉCENTS (que tu
